Network errors do not count as packets, better fonts sizes, colors
- network errors don't go to the queue and are sent with a different signal
- network errors don't mess up the latency chart
- smaller sections headers size
- a bit higher normal font size
- better names for font sizes
- colors for percentages and latency in report
- better quality definitions for the latency conclusion
- bigger integer size for some packets values

+++ pinger-quick/pingdata.h
@@ -21,7 +21,7 @@ public:
     QList<int> *get_packetsQueueTimes();
     float get_lastPacketTime();
     float get_avgTime();
-    QQueue< QPair<int, QString> > get_packets();
+    QQueue<QPair<int, QString>> get_packets();
 
     void resetEverything();
 
@@ -29,13 +29,13 @@ public:
     //QString generateReport(int reportFormat);
 
 private:
-    int pcktLost;
-    int pcktReceived;
-    int pcktSent;
+    unsigned long pcktLost;
+    unsigned long pcktReceived;
+    unsigned long pcktSent;
     int packetsQueueSize;
     float avgTime;
     float lastPacketTime;
-    int totalTime;
+    unsigned long totalTime;
     QQueue<QPair<int, QString>> packetsQueue;
 
     void resizePacketsQueue(int newQueueSize);
